+++ src/component/_generated/api.ts
@@ -10,6 +10,7 @@
 
 import type * as mutations from "../mutations.js";
 import type * as queries from "../queries.js";
+import type * as validation from "../validation.js";
 import type * as validators from "../validators.js";
 
 import type {
@@ -22,14 +23,31 @@ import { anyApi, componentsGeneric } from "convex/server";
 const fullApi: ApiFromModules<{
   mutations: typeof mutations;
   queries: typeof queries;
+  validation: typeof validation;
   validators: typeof validators;
 }> = anyApi as any;
 
+/**
+ * A utility for referencing Convex functions in your app's public API.
+ *
+ * Usage:
+ * ```js
+ * const myFunctionReference = api.myModule.myFunction;
+ * ```
+ */
 export const api: FilterApi<
   typeof fullApi,
   FunctionReference<any, "public">
 > = anyApi as any;
 
+/**
+ * A utility for referencing Convex functions in your app's internal API.
+ *
+ * Usage:
+ * ```js
+ * const myFunctionReference = internal.myModule.myFunction;
+ * ```
+ */
 export const internal: FilterApi<
   typeof fullApi,
   FunctionReference<any, "internal">